+++ src/main/kotlin/com/loadingbyte/cinecred/ui/view/welcome/WelcomeFrame.kt
@@ -97,6 +97,8 @@ class WelcomeFrame(private val welcomeCtrl: WelcomeCtrlComms) : CcFrame(l10n("ui
         panel.preferencesPanel.startPreferencesForm.preferences_start_setProjectHintTrackPending(pending)
     override fun preferences_start_setAppleScriptFileChooser(use: Boolean) =
         panel.preferencesPanel.startPreferencesForm.preferences_start_setAppleScriptFileChooser(use)
+    override fun preferences_start_setTapePreviewResolution(resolution: Int) =
+        panel.preferencesPanel.startPreferencesForm.preferences_start_setTapePreviewResolution(resolution)
     override fun preferences_start_setAccounts(accounts: List<Account>) =
         panel.preferencesPanel.preferences_start_setAccounts(accounts)
     override fun preferences_start_setAccountRemovalLocked(account: Account, locked: Boolean) =